refactor(python): move AgentStatusResolver from core/ to services/agents/
Fixes the import-linter four-tier-layers contract: nexus.core was
importing nexus.services.agents.agent_registry through
core/agent_status_resolver.py's TYPE_CHECKING block. The Python
AgentStatusResolver consumes the Python AgentRegistry, so they belong
in the same tier — services/agents/.

Pure relocation: file move + import updates at 3 sites
(factory/orchestrator.py registration site, the test file's own import,
and the contracts/vfs_paths.py docstring reference). No behavioral
change. The Rust kernel-side AgentStatusResolver
(rust/kernel/src/agent_status_resolver.rs) stays where it is — it
implements a kernel-internal PathResolver trait and is genuinely
kernel-tier.
